10 oct 2023

+++ backend/routes/posts.js
@@ -32,8 +32,11 @@ router.post(
     const post = new Post({
         title : req.body.title,
         content : req.body.content,
-        imagePath : url + "/images/" + req.file.filename
+        imagePath : url + "/images/" + req.file.filename,
+        creator : req.userData.userId
     });
+    //console.log(req.userData);
+    //return res.status(200).json({});
     console.log(post);
     post.save().then(createdPost=>{
         res.status(201).json({
@@ -62,12 +65,17 @@ router.put(
         _id : req.body.id,
         title: req.body.title,
         content : req.body.content,
-        imagePath : imagePath
+        imagePath : imagePath,
+        creator : req.userData.userId
     });
     console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        console.log(result);
-        res.status(200).json({message : 'Update Successful!'});
+    Post.updateOne({_id: req.params.id, creator : req.userData.userId}, post).then(result => {
+        //console.log(result);
+        if(result.modifiedCount > 0){
+            res.status(200).json({message : 'Update Successful!'});
+        }else{
+            res.status(401).json({message : 'Not Authorized!'});
+        }
     })
 });
 
@@ -106,9 +114,14 @@ router.get("", (req, res, next) => {
     });
   });
   router.delete("/:id",checkAuth, (req,res,next)=>{
-    Post.deleteOne({_id : req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({message : 'Post Deleted!'});
+    Post.deleteOne({_id : req.params.id, creator : req.userData.userId}).then(result => {
+        //console.log(result);
+        //res.status(200).json({message : 'Post Deleted!'});
+        if(result.modifiedCount > 0){
+            res.status(200).json({message : 'Deletion Successful!'});
+        }else{
+            res.status(401).json({message : 'Not Authorized!'});
+        }
     })
     console.log(req.params.id);
